+++ 7SIM/.github/scripts/post-build-with-gradle.sh
@@ -17,9 +17,9 @@ echo "ARTIFACTS_DIR=$artifacts_dir" >> "$GITHUB_ENV"
 
 # Move artifacts into a dedicated directory
 mv -v build/reports/kover/htmlDebug/ "$artifacts_dir/koverDebug" || true
-mv -v build/reports/tests/testDebugUnitTest/ "$artifacts_dir" || true
+mv -v build/reports/tests/test{Debug,Release}UnitTest/ "$artifacts_dir" || true
 mv -v build/reports/androidTests/connected/debug/ "$artifacts_dir/androidTestsDebug" || true
-mv -v build/reports/lint-results-debug.html "$artifacts_dir" || true
+mv -v build/reports/lint-results-debug.* "$artifacts_dir" || true
 if ls build/{outputs,reports}/roborazzi >/dev/null; then
     roborazzi_dir="$artifacts_dir/roborazzi"
     mkdir -vp "$artifacts_dir/tests/test" && mv -v tests/test/roborazzi/ "$_"